+++ src/app/common/rootReducer.ts
@@ -1,12 +1,12 @@
 import { combineReducers } from "redux";
-import authReducer from "../Login/auth.reducer";
+import loginReducer from "../Login/Login.reducer";
 /* import slices of state here */
 
 /* Place the object keys for state here followed by the reducer taken from that slice
     e.g. key : keyReducer
 */
 const rootReducer = combineReducers({
-	auth: authReducer,
+	login: loginReducer,
 });
 
 export default rootReducer;
